Add expressions for pricelist assignment
Former-commit-id: a22cd815f673b7bfe46bcb0ad4d39d726b76b946

+++ PLATFORM/Modules/DynamicExpression/VirtoCommerce.DynamicExpressionModule.Data/Promotion/Conditions/CartConditions/ConditionAtNumItemsOfEntryAreInCart.cs
@@ -8,7 +8,7 @@ namespace VirtoCommerce.DynamicExpressionModule.Data.Promotion
 	//[] [] items of entry are in shopping cart
 	public class ConditionAtNumItemsOfEntryAreInCart : ConditionBase, IConditionExpression
 	{
-		public decimal NumItem { get; set; }
+		public int NumItem { get; set; }
 		public bool Exactly { get; set; }
 		public string ProductId { get; set; }
 		public string ProductName { get; set; }
